refactor(router-lite): routing-trigger
